ai_service/requirements.txt
# Resume Builder AI Service - Python Dependencies
# For Render.com deployment - requires Python 3.11

# Web framework
fastapi>=0.100.0
uvicorn[standard]>=0.23.0

# Google Gemini LLM (direct, no CrewAI for simpler deployment)
google-generativeai>=0.3.0

# PDF processing
PyPDF2>=3.0.0
python-multipart>=0.0.6

# Environment
python-dotenv>=1.0.0
pydantic>=2.0.0
